Change some field types, reduce memory usage

+++ PKHeX.Core/Legality/Encounters/EncounterStatic/DreamWorldEntry.cs
@@ -6,7 +6,7 @@ namespace PKHeX.Core
     /// <summary>
     /// Intermediary Representation of Dream World Data
     /// </summary>
-    internal record DreamWorldEntry(int Species, int Level, int Move1 = 0, int Move2 = 0, int Move3 = 0, int Form = 0, int Gender = -1)
+    internal record DreamWorldEntry(int Species, int Level, ushort Move1 = 0, ushort Move2 = 0, ushort Move3 = 0, byte Form = 0, sbyte Gender = -1)
     {
         private int EntryCount => Move1 == 0 ? 1 : Move2 == 0 ? 1 : Move3 == 0 ? 2 : 3;
