+++ PKHeX.Core/Legality/Encounters/EncounterTrade/EncounterTrade2.cs
@@ -11,7 +11,7 @@ namespace PKHeX.Core
         public override int Generation => 2;
         public override int Location => Locations.LinkTrade2NPC;
 
-        public EncounterTrade2(int species, int level, int tid) : base(species, level, GameVersion.GSC)
+        public EncounterTrade2(int species, int level, ushort tid) : base(species, level, GameVersion.GSC)
         {
             TID = tid;
         }
